requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy>=1.24.3
scikit-learn>=1.3.0
Flask==2.3.3
Pillow==10.0.1
gunicorn==21.2.0
